feat: refactor AnswerComment and QuestionComment to extend Comment class
