+++ src/domain/forum/enterprise/entities/question-comment.ts
@@ -1,39 +1,17 @@
-import { Entity } from '@/core/entities/entity'
 import { UniqueEntityId } from '@/core/entities/value-objects/unique-entity-id'
 import { Optional } from '@/core/types/optional'
+import {
+  Comment,
+  CommentProps,
+} from '@/domain/forum/enterprise/entities/comment'
 
-export interface QuestionCommentProps {
-  authorId: UniqueEntityId
-  answerId: UniqueEntityId
-  content: string
-  createdAt: Date
-  updatedAt?: Date
+export interface QuestionCommentProps extends CommentProps {
+  questionId: UniqueEntityId
 }
 
-export class QuestionComment extends Entity<QuestionCommentProps> {
-  get authorId() {
-    return this.props.authorId
-  }
-
-  get content() {
-    return this.props.content
-  }
-
-  get createdAt() {
-    return this.props.createdAt
-  }
-
-  get updatedAt() {
-    return this.props.updatedAt
-  }
-
-  private touch() {
-    this.props.updatedAt = new Date()
-  }
-
-  set content(content: string) {
-    this.props.content = content
-    this.touch()
+export class QuestionComment extends Comment<QuestionCommentProps> {
+  get questionId() {
+    return this.props.questionId
   }
 
   static create(
